+++ mjolnir/omp/UnlimitedGridCellList.hpp
@@ -312,10 +312,10 @@ class UnlimitedGridCellList<OpenMPSimulatorTraits<realT, boundaryT>, potentialT>
 } // mjolnir
 
 #ifdef MJOLNIR_SEPARATE_BUILD
-#include <mjolnir/potential/global/DebyeHuckelPotential.hpp>
-#include <mjolnir/potential/global/ExcludedVolumePotential.hpp>
-#include <mjolnir/potential/global/LennardJonesPotential.hpp>
-#include <mjolnir/potential/global/UniformLennardJonesPotential.hpp>
+#include <mjolnir/forcefield/global/DebyeHuckelPotential.hpp>
+#include <mjolnir/forcefield/global/ExcludedVolumePotential.hpp>
+#include <mjolnir/forcefield/global/LennardJonesPotential.hpp>
+#include <mjolnir/forcefield/global/UniformLennardJonesPotential.hpp>
 
 namespace mjolnir
 {
